Move camayoc.ui.types into camayoc.types.ui
All Camayoc types should be consolidated in single place.
